feat: Add React Intersection Observer package
The package "@radix-ui/react-navigation-menu" has been added to the project dependencies. This package provides a React component for creating navigation menus. It will be used to enhance the navigation functionality of the application.

Note: This commit message follows the conventional commit format, using the type "feat" to indicate a new feature addition.