feat: Update lazy loading configuration for images
The lazy loading configuration for images has been updated in the `next.config.mjs` file. The `remotePatterns` array now includes a new object with the protocol set to 'https' and the hostname set to '**.pakfrag.com'. This change enables lazy loading for images from the 'pakfrag.com' domain, improving the page load performance by loading images only when they are visible on the screen.